requirements.txt
PyYAML~=6.0.2
inotify-simple~=1.3.5
